+++ src/main/java/org/cicirello/search/evo/ShiftedStochasticUniversalSampling.java
@@ -51,9 +51,13 @@ package org.cicirello.search.evo;
  * <p>The runtime to select M population members from a population of size N is
  * O(N + M), which includes the need to generate only a single random double, and O(M) ints.</p>
  *
+ * @deprecated Instead of this class, you should use a combination of {@link FitnessShifter} and 
+ * {@link StochasticUniversalSampling}. This class is scheduled for removal in release 6.0.0.
+ *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
  */
+@Deprecated
 public final class ShiftedStochasticUniversalSampling extends StochasticUniversalSampling {
 	
 	/**