+++ src/main/java/org/cicirello/search/evo/StochasticUniversalSampling.java
@@ -45,7 +45,8 @@ import org.cicirello.math.rand.RandomIndexer;
  *
  * <p><b>This selection operator requires positive fitness values. Behavior is undefined if any 
  * fitness values are less than or equal to 0.</b> If your fitness values may be negative,
- * use {@link ShiftedStochasticUniversalSampling} instead.</p>
+ * you can use {@link FitnessShifter}, which transforms fitness values such that minimum fitness
+ * equals 1.</p>
  *
  * <p>The runtime to select M population members from a population of size N is
  * O(N + M), which includes the need to generate only a single random double, and O(M) random ints.</p>